convert enum impl to use a structured format
